feat(keymap,tui): add key modifiers
It's a bit tricky, but since there are not enough bits for the key
modifiers, the lower three bits of KS_MODIFIER(0b1110 0000) are used
for the key modifiers. Also, add 1 to the key modifiers so that it is
not zero.

+++ src/nvim/api/vim.c
@@ -369,7 +369,7 @@ void nvim_input_mouse(String button, String action, String modifier, Integer gri
     modmask |= mod;
   }
 
-  input_enqueue_mouse(code, (uint8_t)modmask, (int)grid, (int)row, (int)col);
+  input_enqueue_mouse(code, (int)modmask, (int)grid, (int)row, (int)col);
   return;
 
 error:
